requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.54.3
